Extract inline assistant functionality into an `InlineAssistant` global
This will help adding the same functionality in the prompt library.
